Refactor package structure for Velocity implementation.
